Pathfinding optimizations

+++ src/protocolgame.cpp
@@ -789,8 +789,8 @@ void ProtocolGame::parseAutoWalk(NetworkMessage& msg)
 {
 	std::list<Direction> path;
 
-	size_t numdirs = msg.GetByte();
-	for (size_t i = 0; i < numdirs; ++i) {
+	uint8_t numdirs = msg.GetByte();
+	for (uint8_t i = 0; i < numdirs; ++i) {
 		uint8_t rawdir = msg.GetByte();
 		switch (rawdir) {
 			case 1: path.push_back(EAST); break;
